+++ leads/data_persistence/analyzer/__init__.py
@@ -1,2 +1,3 @@
 from leads.data_persistence.analyzer.dp import *
+from leads.data_persistence.analyzer.inference import *
 from leads.data_persistence.analyzer.processor import *
